DELIA-41091: Video freeze when player transition from ads to content
Reason for change: Avoid sending flush start/stop events right after
a flushing seek which pauses basesrc src_loop task and leads to freeze
Test Procedure: tested on pace xi5
Risks: Low
(cherry picked from commit ac19786443894e32881e1063f36a5edd640929d4)

+++ priv_aamp.h
@@ -3087,7 +3087,7 @@ public:
 	 *   @param[in] rate - playback rate
 	 *   @return void
 	 */
-	void FlushStreamSink(double position, int rate);
+	void FlushStreamSink(double position, double rate);
 
 private:
